projects/auth_server/pyproject.toml
[tool.poetry]
name = "simple-msa"
version = "0.1.0"
description = ""
authors = ["pai-cm <cmkim@publicai.co.kr>"]
readme = "README.md"
packages = [
    {include="src"},
    {include='webapp'}
]

[tool.poetry.dependencies]
python = "^3.11,<3.12"
pycryptodome = "^3.20.0"
pyjwt = "^2.8.0"
cryptography = "^42.0.5"
sqlalchemy = "^2.0.29"
aiosqlite = "^0.20.0"
pre-commit = "^3.7.0"
greenlet = "^3.0.3"
fastapi = "^0.110.1"
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"

[tool.pytest.ini_options]
minversion = "6.0"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
